requirements.txt
python-telegram-bot
cfscrape
requests
js2py
transmissionrpc
#google_speech
pytg
